refactor: seprate ext js file and organize them
